requirements.txt
# Core Dependencies
fastapi>=0.68.0,<0.69.0  # Using version compatible with Pydantic v1
uvicorn[standard]>=0.22.0,<0.23.0
python-multipart>=0.0.5,<0.1.0
python-dotenv>=0.19.0,<0.20.0
python-dateutil>=2.8.2,<3.0.0

# Database
sqlalchemy[asyncio]>=1.4.0,<2.0.0  # Using v1 for compatibility
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
aiosqlite>=0.19.0,<0.20.0

# Pydantic
pydantic[email]>=1.10.0,<2.0.0  # Using Pydantic v1 for FastAPI compatibility
pyyaml>=6.0.0,<7.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0  # For file uploads
bcrypt>=4.0.0,<5.0.0  # For password hashing

# Async
anyio>=3.0.0,<4.0.0
httpx==0.25.0  # Pinned for compatibility with test dependencies
asgiref>=3.6.0,<4.0.0  # ASGI support

# Utils
python-slugify>=5.0.0,<6.0.0
structlog>=23.1.0,<24.0.0
typing-extensions>=4.0.0,<5.0.0

# Scheduler
APScheduler>=3.10.0,<4.0.0
tzlocal>=4.3.0,<5.0.0  # Timezone support

# Testing
# Note: Test dependencies are now in requirements-dev.txt

# Notifications
twilio>=8.0.0,<9.0.0
python-telegram-bot>=20.0,<21.0
# Note: For development, install with: pip install -r requirements-dev.txt
